+++ pom.xml
@@ -19,6 +19,11 @@
 	</properties>
 	<dependencies>
 	    <dependency>
+	        <groupId>org.thymeleaf.extras</groupId>
+	        <artifactId>thymeleaf-extras-java8time</artifactId>
+	        <version>3.0.4.RELEASE</version>
+	    </dependency>
+	    <dependency>
 	        <groupId>org.springframework.boot</groupId>
 	        <artifactId>spring-boot-starter-thymeleaf</artifactId>
 	        <version>2.7.0</version>
