+++ src/main/java/com/teama/bioskop/RestControllers/FilmsController.java
@@ -1,4 +1,4 @@
-package com.teama.bioskop.Controllers;
+package com.teama.bioskop.RestControllers;
 
 import com.teama.bioskop.DTOs.FilmRequestDTO;
 import com.teama.bioskop.DTOs.FilmResponseDTO;
