+++ src/main/resources/application.properties
@@ -5,3 +5,5 @@ spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.PostgreSQLDialect
 spring.jpa.properties.hibernate.jdbc.lob.non_contextual_creation=true
 spring.jpa.hibernate.ddl-auto=none
 server.port=8082
+
+spring.mvc.hiddenmethod.filter.enabled=true
