+++ src/main/resources/templates/homepage.html
@@ -1,0 +1,17 @@
+<!DOCTYPE html>
+<html xmlns:th="http://www.thymeleaf.org">
+<head>
+    <meta charset="UTF-8">
+    <meta http-equiv="X-UA-Compatible" content="IE=edge">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Document</title>
+</head>
+<body>
+    <ul th:each="test : ${tests}">
+        <li th:text="${test.userId}"></li>
+        <li th:text="${test.username}"></li>
+        <li th:text="${test.emailAddress}"></li>
+        <li th:text="${test.password}"></li>
+    </ul>
+</body>
+</html>
